Rag/app/config.py
import os
from dotenv import load_dotenv
load_dotenv()

GEMINI_API_KEY = os.getenv("GEMINI_API_KEY", "")
OPENWEATHER_API_KEY = os.getenv("OPENWEATHER_API_KEY", "")
QDRANT_URL = os.getenv("QDRANT_URL", "http://localhost:6333")
QDRANT_API_KEY = os.getenv("QDRANT_API_KEY", "")
PDFS_DIR = os.getenv("PDFS_DIR", "pdfs")
LANGSMITH_API_KEY = os.getenv("LANGSMITH_API_KEY", "")
