requirements.txt
google-generativeai
qdrant-client
langchain
langgraph
streamlit
pypdf
python-dotenv
requests
pytest
tqdm

